Split E2E tests into 2, clean up fixtures based off feedback.
