+++ .github/data/matrix-smoke-oss.json
@@ -36,38 +36,45 @@
       "platforms": "linux/arm64, linux/amd64"
     },
     {
-      "label": "policies 1/5",
+      "label": "policies 1/6",
       "image": "alpine",
       "type": "oss",
       "marker": "'policies and not policies_rl and not policies_ac and not policies_jwt and not policies_mtls and not policies_cache and not policies_cors and not policies_external_auth'",
       "platforms": "linux/arm64, linux/amd64"
     },
     {
-      "label": "policies 2/5",
+      "label": "policies 2/6",
       "image": "alpine",
       "type": "oss",
       "marker": "'policies_ac or policies_jwt'",
       "platforms": "linux/arm64, linux/amd64"
     },
     {
-      "label": "policies 3/5",
+      "label": "policies 3/6",
       "image": "alpine",
       "type": "oss",
       "marker": "'policies_rl or policies_cors'",
       "platforms": "linux/arm64, linux/amd64"
     },
     {
-      "label": "policies 4/5",
+      "label": "policies 4/6",
       "image": "alpine",
       "type": "oss",
       "marker": "'policies_mtls or policies_cache or otel'",
       "platforms": "linux/arm64, linux/amd64"
     },
     {
-      "label": "policies 5/5",
+      "label": "policies 5/6",
       "image": "alpine",
       "type": "oss",
-      "marker": "'policies_external_auth'",
+      "marker": "'policies_external_auth and not policies_external_auth_tls'",
+      "platforms": "linux/arm64, linux/amd64"
+    },
+    {
+      "label": "policies 6/6",
+      "image": "alpine",
+      "type": "oss",
+      "marker": "'policies_external_auth_tls'",
       "platforms": "linux/arm64, linux/amd64"
     },
     {
