+++ pyproject.toml
@@ -54,6 +54,7 @@ markers = [
     "otel",
     "policies",
     "policies_external_auth",
+    "policies_external_auth_tls",
     "policies_rl",
     "policies_rl_vs",
     "policies_rl_vsr",
