+++ tests/conftest.py
@@ -16,7 +16,11 @@ from settings import (
     NS_COUNT,
     NUM_REPLICAS,
 )
-from suite.utils.resources_utils import are_all_pods_in_ready_state, get_first_pod_name, wait_before_test
+from suite.utils.resources_utils import (
+    are_all_pods_in_ready_state,
+    get_first_pod_name,
+    wait_before_test,
+)
 
 
 def pytest_addoption(parser) -> None:
@@ -141,7 +145,12 @@ def pytest_addoption(parser) -> None:
 
 
 # import fixtures into pytest global namespace
-pytest_plugins = ["suite.fixtures.fixtures", "suite.fixtures.ic_fixtures", "suite.fixtures.custom_resource_fixtures"]
+pytest_plugins = [
+    "suite.fixtures.fixtures",
+    "suite.fixtures.ic_fixtures",
+    "suite.fixtures.custom_resource_fixtures",
+    "suite.utils.external_auth_utils",
+]
 
 
 def pytest_configure(config):
